chore: test and document further issues with @opentelemetry / mysql2
Added a section to the README with instructions on how to checkout
the fixed mysql2 version and test it with this project.

However, I noticed further problems (the fixed version is not
instrumented anymore. This would require further patches.
I tested these patches and described them. With the
further patches, instrumentation is actually
working).